src/requirements.txt
requests~=2.32.3
boto3~=1.35.86
python-dotenv~=1.0.1
python-telegram-bot~=21.9